move shared code into a common lib
