+++ common/src/main/java/my/raptive/url/common/repository/ProcessInfoRepository.java
@@ -1,6 +1,6 @@
-package my.raptive.url.rest.repository;
+package my.raptive.url.common.repository;
 
-import my.raptive.url.rest.rs.WebContentDto;
+import my.raptive.url.common.rs.WebContentDto;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
@@ -16,7 +16,7 @@ public interface ProcessInfoRepository extends JpaRepository<ProcessInfo, Long> 
         @Query("SELECT p FROM ProcessInfo p join Url u on p.urlId = u.id WHERE u.id = :urlId")
         ProcessInfo findByUrlId(@Param("urlId") long urlId);
 
-        @Query("SELECT new my.raptive.url.rest.rs.WebContentDto(u.url, p.title, p.description, p.body, p.errorMessage, u.processStatus) " +
+        @Query("SELECT new my.raptive.url.common.rs.WebContentDto(u.url, p.title, p.description, p.body, p.errorMessage, u.processStatus) " +
                 "FROM ProcessInfo p join Url u on p.urlId = u.id " +
                 "join Request r " +
                 "WHERE r.id = :requestId")
