+++ common/src/main/java/my/raptive/url/common/repository/Request.java
@@ -1,4 +1,4 @@
-package my.raptive.url.processor.repository;
+package my.raptive.url.common.repository;
 
 import jakarta.persistence.*;
 import lombok.Getter;
