+++ common/src/main/java/my/raptive/url/common/repository/RequestRepository.java
@@ -1,4 +1,4 @@
-package my.raptive.url.processor.repository;
+package my.raptive.url.common.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
