+++ common/src/main/java/my/raptive/url/common/repository/ShardCountRepository.java
@@ -1,7 +1,6 @@
-package my.raptive.url.rest.repository;
+package my.raptive.url.common.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
