+++ pom.xml
@@ -12,6 +12,7 @@
 		<relativePath/> <!-- lookup parent from repository -->
 	</parent>
 	<modules>
+		<module>common</module>
 		<module>rest</module>
 		<module>processor</module>
 	</modules>
