--- rest/src/main/java/my/raptive/url/rest/repository/RequestRepository.java
@@ -1,14 +1,0 @@
-package my.raptive.url.rest.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.stereotype.Repository;
-import org.springframework.transaction.annotation.Transactional;
-
-@Transactional
-@Repository
-public interface RequestRepository extends JpaRepository<Request, Long> {
-
-    @Query("SELECT r FROM Request r WHERE r.id = ?1")
-    Request findByRequestId(long requestId);
-}
