--- rest/src/main/java/my/raptive/url/rest/repository/ShardCount.java
@@ -1,18 +1,0 @@
-package my.raptive.url.rest.repository;
-
-import jakarta.persistence.*;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import lombok.experimental.Accessors;
-
-@Getter
-@Setter
-@Accessors(chain = true)
-@NoArgsConstructor
-@Entity
-public class ShardCount {
-
-    @Id
-    private int shardCount;
-}
